requirements.txt
polars
openai
python-dotenv
kaggle
sounddevice
scipy
langchain
bs4
whisper
pyspark